Update Prisma Schema
Add new Session and Response models. Each Session represents an entire
peerprep session between two users. The response stored are for each
question. As questions are stored in mongodb, a string questionId is
used here instead. Also added a new enum for response status
Note the cascading delete for response objects when the parent session
is deleted

Add Session and Response routers and endpoints

Only a single post endpoint was created for Response since we will only
require to update/insert it. Deletion will cascade when the session is
deleted. For sessions, there is a create, delete and get endpoint. The
get endpoint returns all the sessions, responses and partners of a user.

Add SessionDTO to convert bigint to string